chore: add Class AppError

+++ src/storage/group/groupCreate.ts
@@ -1,12 +1,19 @@
 import AsyncStorage from "@react-native-async-storage/async-storage";
 import { GROUP_COLLECTION } from "@storage/storageConfig";
 import { groupsGetAll } from "./groupsGetAll";
+import { AppError } from "@utils/AppError";
 
 export async function groupCreate(newGroupName: string) {
     try {
-        const storageGroups = await groupsGetAll();
+        const storedGroups = await groupsGetAll();
 
-        const storage = JSON.stringify([...storageGroups,  newGroupName]);
+        const groupAllReadyExists = storedGroups.includes(newGroupName);
+
+        if (groupAllReadyExists) {
+            throw new AppError('Já existe um grupo cadastro com esse nome!');
+        }
+
+        const storage = JSON.stringify([...storedGroups,  newGroupName]);
 
         await AsyncStorage.setItem(GROUP_COLLECTION, storage);
     } catch(error) {
